check if a guild exists to even check muted role

+++ bot/listen.js
@@ -38,11 +38,14 @@ bot.on('message', async function (message) {
     /**
      * Listen to messages and convert into params
      */
-    let muted = message.member.roles.find(function (r) {
-        if (r.name.toLowerCase().includes('mute')) return r
-    })
-    if (muted)
-        message.delete();
+    if(message.guild)
+    {
+        let muted = message.member.roles.find(function (r) {
+            if (r.name.toLowerCase().includes('mute')) return r
+        })
+        if (muted)
+            message.delete();
+    }
     if (message.content.startsWith(settings.identifier)) {
         /**Extracting params */
         let params = message.content.substring(settings.identifier.length).trim();
